lwt: upgrade most recently committed mutation to latest schema during prepare
Currently it is upgraded during learn on a replica. The are two problems
with this.  First the column mapping may not exist on a replica if
it missed this particular schema (because it was down for instance)
and the mapping history is not part of the schema. Second lwt request
coordinator may not have the schema for the most recently committed
mutation as well (because it was freed from the registry already) and
when a replica tries to retrieve the schema from the coordinator the
retrieval will fail causing the whole request to fail.

Both of those problems can be fixed by upgrading most recently committed
mutation during prepare on a node is is stored at. To upgrade the
mutation its column mapping is needed and it is guarantied that it will
be present at the node the mutation is stored at since it is pre-request
to store it that the corresponded schema is available. After that the
mutation is processed using latest schema that will be available on all
nodes.