service: raft: add get_read_apply_mutex to group0_guard
group0_guard has a mutex that is taken in start_operation() before
accessing local group0 state machine's state. The same mutex is taken by
group0's apply function. It is needed to provide mutual exclusion
between a read side and a write side. It is released by group0_client just
before adding an entry with wait_type::applied since otherwise
there will be a deadlock: apply will not complete until the mutex is
released. What it means in practice is that the old state machine's
state is no longer valid immediately after group_client::add_entry()
call.

Sometimes we want to extent validity of the local state for a little
bit longer. For that we need to delay the call to group0::apply()
until current state machine's state is no longer in use. To do that this
patch adds a function that allows to move the mutex units to an external
code and let it release them when it knows that the previous state is
no longer needed. To avoid the deadlock the group0 client's code uses
wait_type::committed to add an entry in case it detects that it cannot
release the mutex by itself.