storage_service: raft topology: rollback topology operation on streaming failure.
Currently if a streaming fails during a topology operation the streaming
is retried until is succeeds. If it will never succeed it will be
retried forever. There is no way to stop the topology operation.

This patch introduce the rollback mechanism on streaming failure. If
streaming fails during bootstrap/replace the bootstrapping/replacing
node is moved to the left state and the operation has to be restarted
after removing data directory. If streaming fails during
decommission/remove the node is moved back to normal and the operation
need to be restarted after the failure reason is eliminated.