storage_service: raft topology: get rid of raft_topology_cmd::shutdown RPC
The shutdown RPC is issued to a decommissioned node in left_token_ring
state to let it know that it may shutdown itself now. This is fine if
the decommission succeeded, but it become a nuisance if it fails.
We need to invent a way to signal the future the shutdown RPC triggers
with an error somehow. Unlike shutdown RPC, this other way should not
fail otherwise decommission operation will never complete. But luckily
we do not need shutdown RPC at all since we can detect that decommission
completed by checking that the decommissioned mode moved to
left_token_ring state. For that it is enough to run simple barrier RPC
to the node.