gossip: remove unused STATUS_MOVING state
Moving operation was removed by 4a0b561376c591026333cd2694b92bad2d8c971d
and since then the state is unused. Even back then it worked only for
the case of one token so it is safe to say we never used it. Lets
remove the remains of the code instead of carrying it forever.