storage_service: topology coordinator: mark nodes as needing cleanup when required
A cleanup needs to run when a node loses an ownership of a range (during
bootstrap) or if a range movement to an normal node failed (removenode,
decommission failure). Mark all node as "cleanup needed" in those cases.
Later it can be optimized to require cleanup only on those nodes that
were actually affected.