storage_service: topology coordinator: wait for cleanup to complete before starting new topology operation
It is unsafe to start new topology operation before cleanup is note done
after the previous one. Wait for all nodes to complete the cleanup
before starting new topology operation. Note that some topology
operation should be safe to start even if cleanup did not complete yet.
but lets leave it for an optimization for now.