storage_service: topology coordinator: do fencing even if draining failed
Token metadata barrier consists for two steps. First old request are
drained and then requests that are not drained are fenced. But currently
if draining fails then fencing is note done. This is fine if the
barrier's failure handled by retrying, but we when to start handling
errors differently. In fact during topology operation rollback we
already do not retry failed barrier.

The patch fixes the metadata barrier to do fencing even if draining
failed.