storage_service: topology coordinator: add rollback_to_normal node state
When a topology coordinator rolls back from unsuccessful topology operation it
advances the fence (which is now in raft state) after moving to normal
state. We do not want this to fail (only majority of nodes is needed for
it to not to), but currently it may fail in case the coordinator moves
to another node after changing the rollback node's state to normal, but
before updating the fence. To solve that the rollback operation needs to
go through a new rollback_to_normal state that will do the fencing
before moving to normal. This patch introduces that state, but not use
it yet.