storage_service: topology_coordinator: use new rollback_to_normaltate state in the rollback procedure
Go through the rollback_to_normaltate to normal state when the node
needs to move to normal during the rollback and update fence in this
state before moving the node to normal. This guaranties that the fence
update will not be missed. Not that when a node moves to left state it
already passes through left_token_ring which guaranties the same.