storage_service: topology coordinator:: make topology coordinator livecycle subscriber
We want to change the coordinator to consider nodes liveness when
processing the topology operation queue. If there is no enough live
nodes to process any of the ops we want to cancel them. For that to work
we need to be able to kick the coordinator if it waits for some event
that prevent it from processing the request queue otherwise.