storage_service: disable current cleanup API if topology coordinator is enabled
I is not safe to start cleanup manually without integration with the
topology coordinator, so disable it. Next patch will provide the
integration.

Note that we leave the API available in non release mode for testing
(since test.py runs with topology coordinator enabled and some tests
want to be able to test old functionality).