group0: make sure that address map has an entry for each entry in a new configuration
ID->IP mapping is added to raft address map when the mapping is first
appear in the gossiper, but it is added as expiring entry. It becomes
non expiring when a node is added to raft configuration. But when a node
joins those two events may be distant in time (since the node's request
may sit in the topology coordinator queue for a while) and a mapping may
expire already from the map. This patch makes sure to transfer the
mapping from the gossiper to the map when it adds a node to the config.